criando teste de integração

+++ LocadoraAutomoveis.Infra.Orm/ModuloPlano/IRepositorioOrmPlano.cs
@@ -6,7 +6,7 @@ namespace LocadoraVeiculos.Infra.BancoDados
 {
     public interface IRepositorioOrmPlano : IRepositorioOrmBase<Plano>
     {
-        public Plano SelecionarPorValor(float valor);
+        public Plano SelecionarPorValor(double valor);
         public Plano SelecionarPorGrupo(string valor);
     }
 }
